fix sim detectorMaps
The detectorMaps coming from the simulator have non-physical features at
either end, possibly due to spline artifacts. Jim Gunn says, 'There is
nothing physical which could cause this.' We suspect the problem is in the
splining of the optical model. This change includes the results of a
workaround which mitigates the problem by dropping deviant points from
the wavelength spline.