Sep05_v1 version of the solutions, which implement work from PIPE2D-777, PIPE2D-778, PIPE2D-872 and PIPE2D-881.
These solutions were created from defocused images for 29 fibers (10 fiber configuration and 19 fibers from 21 fiber configuration).
The biggest difference to previous solutions is that I implemented centering algorithm based on lsst.afw.math.offsetImage function and the downsampling is done with galsim.Lanczos(5) algorithm.
In this way these are consistent with the way that the pipeline is centering and downsampling the upsampled psf-images to match the data.