add makePfsDefects script
This script was originally called ingestPfsDefects and lived in
obs_pfs.

We need this script to work for both Gen2 and Gen3 LSST middleware,
so having it do the ingest for Gen2 only is not appropriate. It's
easiest if we just have the script do the translation, and then
call the ingest appropriate to the middleware separately.

Moved from obs_pfs because this script is only relevant to the
files in this repo. The script used to generate the defects files
in a staging directory, but Gen3 wants the defects that it ingests
to come from this package.