nirDefects: fix header
The LSST v26 stack is having trouble reading the metadata in
the new defects files, but is fine with the older ones; but
the older ones have no defects in them. Moved the defects from
the new files into the old ones, so we have the good metadata
and half the number of files.